refactor(#1369): extract startup_transport_modes inline tests
